Bug fixes and model redesign.
Persisting session info in db correctly now.
Form 2: User selects cuny subjects for the sending and receiving institutions separately, with
the disciplines for various sending/receiving colleges displayed for infomational purposes only.

+++ mysession.py
@@ -24,7 +24,6 @@
       session_dict blob,
       expiration_time float);
 """
-from collections import namedtuple
 import time, datetime
 import pickle
 from uuid import uuid4
@@ -37,13 +36,15 @@ fh = logging.FileHandler('debugging.log')
 formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
 fh.setFormatter(formatter)
 logger.addHandler(fh)
+
 sql_logger = logging.getLogger('sqlite3')
 sql_logger.setLevel(logging.DEBUG)
-# fh = logging.FileHandler('debugging.log')
-# formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
-# fh.setFormatter(formatter)
-sql_logger.addHandler(fh)
+sfh = logging.FileHandler('debugging.log')
+sformatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+sfh.setFormatter(sformatter)
+sql_logger.addHandler(sfh)
 
+sqlite3.enable_callback_tracebacks(True)
 
 class MySession:
 
@@ -66,18 +67,12 @@ class MySession:
       logger.debug('type of session_key is {}'.format(type(session_key)))
       self.session_key = session_key
       logger.debug('self.session_key is now {}'.format(self.session_key))
-      self.cursor.execute("select expiration_time from sessions where session_key = ?",(self.session_key,))
-      logger.debug('self.cursor is {}'.format(self.cursor))
-      row = self.cursor.fetchone()
-      logger.debug('row is {}'.format(row))
-      expiration_time = self.cursor.fetchone()[0]
-      logger.debug('expiration_time is {} and time.time() is {}'.format(expiration_time, time.time()))
-      if expiration_time and expiration_time > time.time():
-        self.initialized = True
+      if self.is_expired(self.session_key):
+        logger.debug('{} is expired: deleting')
+        self.cursor.execute('delete from sessions where session_key = ?', (self.session_key,))
+        raise Exception('Unable to clone: session expired')
       else:
-        # Reference is to an expired session: delete it from db and reinitialize
-        if expiration_time:
-          self.cursor.execute("delete from sessions where session_key = ?", (session_key,))
+        self.initialized = True
     logger.debug('self.initialized is {}'.format(self.initialized))
     if not self.initialized:
       self.cursor.execute("insert into sessions values(?, ?, ?)", (self.session_key,
@@ -86,15 +81,22 @@ class MySession:
     self.connection.commit()
 
   def __str__(self):
+    logger.debug('__str__()')
     self.cursor.execute("select session_dict from sessions where session_key = ?", (self.session_key,))
     self.mydict = pickle.loads(self.cursor.fetchone()[0])
     return 'Mysession[{}] {} keys'.format(self.session_key, len(self.mydict))
 
   def __del__(self):
-    self.cursor.execute("delete from sessions where session_key = ?", (self.session_key,))
-    self.connection.commit()
+    logger.debug('__del__()')
+    if self.is_expired(self.session_key):
+      self.cursor.execute("delete from sessions where session_key = ?", (self.session_key,))
+      self.connection.commit()
+      logger.debug('Session expired: deleted')
+    else:
+      logger.debug('Session not expired: not deleted')
 
   def __setitem__(self, key, value):
+    logger.debug('__setitem__({}, {})'.format(key, value))
     self.cursor.execute("select session_dict from sessions where session_key = ?", (self.session_key,))
     self.mydict = pickle.loads(self.cursor.fetchone()[0])
     self.mydict[key] = value
@@ -103,19 +105,31 @@ class MySession:
     self.connection.commit()
 
   def __getitem__(self, key):
+    logger.debug('__getitem__({})'.format(key))
     self.cursor.execute("select session_dict from sessions where session_key = ?", (self.session_key,))
     self.mydict = pickle.loads(self.cursor.fetchone()[0])
     return self.mydict[key] # KeyError if key not in session
 
   def __len__(self):
+    logger.debug('__len__()')
     self.cursor.execute("select session_dict from sessions where session_key = ?", (self.session_key,))
     self.mydict = pickle.loads(self.cursor.fetchone()[0])
     return len(self.mydict)
 
   def __bool__(self):
+    logger.debug('__bool__()')
     return True
 
   def keys(self):
+    logger.debug('keys()')
     self.cursor.execute("select session_dict from sessions where session_key = ?", (self.session_key,))
     self.mydict = pickle.loads(self.cursor.fetchone()[0])
     return [key for key in self.mydict]
+
+  def is_expired(self, session_key):
+    self.cursor.execute("select expiration_time from sessions where session_key = ?",(session_key,))
+    row = self.cursor.fetchone()
+    if row == None or len(row) == 0: return True
+    expiration_time = row[0]
+    if expiration_time != None and expiration_time > time.time(): return False
+    return True
